Fixed unable to reach the specified max level in XPTable
Fixed the AP cost display issue in NPC combat control mod.
Added a tweak for weapon perk modifier to AIBestWeaponFix.
Minor edits to code/documents.
Updated version number.

+++ sfall/Modules/Worldmap.cpp
@@ -443,7 +443,7 @@ void StartingStatePatches() {
 	date = GetConfigInt("Misc", "StartDay", -1);
 	if (date >= 0 && date < 31) {
 		dlog("Applying starting day patch.", DL_INIT);
-		SafeWrite8(0x4A3356, date);
+		SafeWrite8(0x4A3356, static_cast<BYTE>(date));
 		dlogr(" Done", DL_INIT);
 	}
